device/requirements.txt
awscrt==0.16.14
awsiotsdk==1.13.0
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
requests==2.28.2
urllib3==1.26.15
